Conference List UI Component  (#77)
* began conference list

* Conferences hideable and conference component functional

* Style fixes.

* Fixed rare race condition. -> slower :/

* Conference Component, Ticket reference BBB meta, attendence list.

* Show hidden conference if current user is attending.

+++ src/main/kotlin/de/thm/mni/ii/classroom/services/ClassroomUserService.kt
@@ -5,11 +5,12 @@ import de.thm.mni.ii.classroom.event.TicketEvent
 import de.thm.mni.ii.classroom.event.UserAction
 import de.thm.mni.ii.classroom.event.UserEvent
 import de.thm.mni.ii.classroom.model.classroom.ClassroomInfo
+import de.thm.mni.ii.classroom.model.classroom.Conference
 import de.thm.mni.ii.classroom.model.classroom.ConferenceInfo
+import de.thm.mni.ii.classroom.model.classroom.DigitalClassroom
 import de.thm.mni.ii.classroom.model.classroom.Ticket
 import de.thm.mni.ii.classroom.model.classroom.User
 import de.thm.mni.ii.classroom.model.classroom.UserCredentials
-import de.thm.mni.ii.classroom.security.exception.UnauthorizedException
 import de.thm.mni.ii.classroom.util.component1
 import de.thm.mni.ii.classroom.util.component2
 import org.slf4j.LoggerFactory
@@ -23,20 +24,21 @@ import java.lang.IllegalArgumentException
 @Service
 class ClassroomUserService(
     private val classroomInstanceService: ClassroomInstanceService,
-    private val conferenceService: ConferenceService
+    private val conferenceService: ConferenceService,
 ) {
 
     private val logger = LoggerFactory.getLogger(ClassroomUserService::class.java)
 
     fun userConnected(userCredentials: UserCredentials, socketRequester: RSocketRequester): Mono<Void> {
         return classroomInstanceService.getClassroomInstance(userCredentials.classroomId)
-            .doOnNext { classroom -> this.conferenceService.updateConferences(classroom).subscribe() }
-            .zipWhen { classroom ->
+            .doOnNext { classroom ->
+                this.classroomInstanceService.abortDeletion(classroom)
+            }.zipWhen { classroom ->
                 classroom.connectSocket(userCredentials, socketRequester)
             }.delayUntil { (classroom, userDisplay) ->
                 classroom.sendToAll(UserEvent(userDisplay, UserAction.JOIN))
             }.doOnSuccess {
-                logger.info("$userCredentials connected to ${userCredentials.classroomId}!")
+                logger.debug("$userCredentials connected to ${userCredentials.classroomId}!")
             }.map {
                 socketRequester.rsocket()!!
             }.doOnNext { socket ->
@@ -56,9 +58,13 @@ class ClassroomUserService(
                 classroom.sendToAll(UserEvent(user, userAction = UserAction.LEAVE))
             }.doOnNext {
                 if (throwable == null) {
-                    logger.info("$userCredentials disconnected from ${userCredentials.classroomId}!")
+                    logger.debug("$userCredentials disconnected from ${userCredentials.classroomId}!")
                 } else {
                     logger.error("$userCredentials disconnected from ${userCredentials.classroomId} with error {}!", throwable.message)
+                }
+            }.doOnNext { (classroom, _) ->
+                if (!classroom.hasUserJoined()) {
+                    this.classroomInstanceService.scheduleClassroomDeletion(classroom)
                 }
             }.subscribe()
     }
@@ -85,46 +91,60 @@ class ClassroomUserService(
             }.subscribe()
     }
 
-    fun assignTicket(userCredentials: UserCredentials, receivedTicket: Ticket) {
+    fun updateTicket(userCredentials: UserCredentials, receivedTicket: Ticket) {
+        if (!userCredentials.isPrivileged()) {
+            logger.warn(
+                "User ${userCredentials.fullName} is not authorized " +
+                    "to change ticket #${receivedTicket.ticketId}!"
+            )
+            return
+        } else if (receivedTicket.assignee != null && !receivedTicket.assignee!!.isPrivileged()) {
+            logger.warn("User ${receivedTicket.assignee} may not be assigned to a ticket!")
+            return
+        }
         classroomInstanceService
             .getClassroomInstance(userCredentials.classroomId)
-            .filter {
-                userCredentials.isPrivileged() && receivedTicket.assignee!!.isPrivileged() && receivedTicket.classroomId == userCredentials.classroomId
-            }.switchIfEmpty {
-                Mono.error(UnauthorizedException("User not authorized to assign ticket!"))
-            }.flatMap {
-                it.assignTicket(receivedTicket, receivedTicket.assignee!!)
-            }.delayUntil { (ticket, classroom) ->
-                classroom.sendToAll(TicketEvent(ticket, TicketAction.ASSIGN))
-            }.doOnSuccess { (ticket, classroom) ->
+            .zipWhen { classroom ->
+                classroom.updateTicket(receivedTicket)
+            }.delayUntil { (classroom, ticket) ->
+                classroom.sendToAll(TicketEvent(ticket, TicketAction.UPDATE))
+            }.doOnSuccess { (classroom, ticket) ->
                 logger.info("Ticket ${classroom.classroomName} / ${ticket.ticketId} assigned to ${ticket.assignee!!.fullName}!")
             }.subscribe()
     }
 
     fun closeTicket(userCredentials: UserCredentials, ticket: Ticket) {
-        classroomInstanceService
-            .getClassroomInstance(userCredentials.classroomId)
-            .filter {
-                ticket.classroomId == userCredentials.classroomId &&
-                    (userCredentials.isPrivileged() || ticket.creator == userCredentials)
-            }.switchIfEmpty {
-                Mono.error(UnauthorizedException("User not authorized to delete ticket!"))
-            }.flatMap { classroom ->
+        if (!userCredentials.isPrivileged() && ticket.creator != userCredentials) {
+            logger.warn("User ${userCredentials.fullName} not authorized to delete ticket #${ticket.ticketId}!")
+            return
+        }
+        classroomInstanceService.getClassroomInstance(userCredentials.classroomId)
+            .delayUntil { classroom ->
                 classroom.deleteTicket(ticket)
-            }.delayUntil { (ticket, classroom) ->
+            }.delayUntil { classroom ->
                 classroom.sendToAll(TicketEvent(ticket, TicketAction.CLOSE))
-            }.doOnSuccess { (ticket, classroom) ->
+            }.flatMap { classroom ->
                 logger.info("Ticket ${classroom.classroomName} / ${ticket.ticketId} assigned to ${ticket.assignee?.fullName ?: "N/A"}!")
+                closeConferenceOfClosedTicket(ticket, classroom)
             }.subscribe()
     }
 
-    fun getUsers(userCredentials: UserCredentials): Flux<UserCredentials> {
-        return classroomInstanceService
-            .getClassroomInstance(userCredentials.classroomId)
-            .flatMapMany { it.getUsersFlux() }
+    private fun closeConferenceOfClosedTicket(ticket: Ticket, classroom: DigitalClassroom): Mono<Void> {
+        return classroom.conferences.getConferenceOfTicket(ticket.conferenceId)
+            .filter { conference ->
+                conference.attendees.isEmpty()
+            }.doOnNext { conference ->
+                this.conferenceService.scheduleConferenceDeletion(classroom, conference, 10)
+            }.then()
     }
 
-    fun getUserDisplays(userCredentials: UserCredentials): Flux<User> {
+    fun getOfflineUsers(userCredentials: UserCredentials): Flux<User> {
+        return classroomInstanceService
+            .getClassroomInstance(userCredentials.classroomId)
+            .flatMapMany { it.getOfflineUsers() }
+    }
+
+    fun getUsers(userCredentials: UserCredentials): Flux<User> {
         return classroomInstanceService
             .getClassroomInstance(userCredentials.classroomId)
             .flatMapMany { it.getUsersFlux() }
@@ -133,7 +153,7 @@ class ClassroomUserService(
     fun getClassroomInfo(userCredentials: UserCredentials): Mono<ClassroomInfo> {
         return classroomInstanceService
             .getClassroomInstance(userCredentials.classroomId)
-            .cast(ClassroomInfo::class.java)
+            .map(DigitalClassroom::getClassroomInfo)
     }
 
     fun getConferences(userCredentials: UserCredentials): Flux<ConferenceInfo> {
@@ -141,7 +161,7 @@ class ClassroomUserService(
             .getClassroomInstance(userCredentials.classroomId)
             .flatMapMany { classroom ->
                 classroom.conferences.getConferences()
-            }.map(::ConferenceInfo)
+            }.map(Conference::toConferenceInfo)
     }
 
     fun changeVisibility(userCredentials: UserCredentials, event: UserEvent) {
